+++ src/api/Category.js
@@ -3,9 +3,8 @@ const router = express.Router();
 const cateogry = require("../models/Category");
 const multer = require("multer");
 const path = require("path");
-
-const { tokengenerate } = require("../middleware/auth");
 const category = require("../models/Category");
+const fs = require("fs");
 
 const storage = multer.diskStorage({
   destination: function (req, file, cb) {
@@ -33,16 +32,12 @@ router.post("/", upload.array("file"), (req, res) => {
         .status(200)
         .send({ message: "All input is required", success: false });
     } else {
-      const date = new Date();
-      req.body.created_at = date.toLocaleString();
-
       const Category = new cateogry(req.body);
       Category.save().then((item) => {
         res.status(200).send({
           message: "Data save into Database",
           data: item,
           success: true,
-          token: tokengenerate({ user: req.user }),
         });
       });
     }
@@ -50,9 +45,9 @@ router.post("/", upload.array("file"), (req, res) => {
     res.status(400).json({ message: err.message, success: false });
   }
 });
-router.put("/", upload.array("file"), (req, res) => {
+router.put("/:id", upload.array("file"), (req, res) => {
   try {
-    const { id } = req.query;
+    const { id } = req.params;
 
     if (req.files) {
       req.body.image = req.files[0] ? req.files[0].path : "";
@@ -60,9 +55,6 @@ router.put("/", upload.array("file"), (req, res) => {
     if (!id) {
       res.status(200).send({ message: "id is not specify", success: false });
     } else {
-      const date = new Date();
-      req.body.updated_at = date.toLocaleString();
-
       cateogry.updateOne({ _id: id }, req.body, (err, result) => {
         if (err) {
           res.status(200).send({ message: err.message, success: false });
@@ -71,7 +63,6 @@ router.put("/", upload.array("file"), (req, res) => {
             message: "Data updated Successfully",
             success: true,
             data: result,
-            token: tokengenerate({ user: req.user }),
           });
         }
       });
@@ -86,16 +77,22 @@ router.delete("/", (req, res) => {
     if (!id) {
       res.status(200).send({ message: "id is not specify", success: false });
     } else {
-      cateogry.deleteOne({ _id: id }, (err, result) => {
-        if (!result) {
-          res.status(200).send({ message: err.message, success: false });
-        } else {
-          res.status(200).send({
-            message: "Data deleted Successfully",
-            success: true,
-            data: result,
-            token: tokengenerate({ user: req.user }),
+      cateogry.findeOne({ _id: id }, (err, result) => {
+        if (result) {
+          fs.unlink(result.image, () => {});
+          cateogry.deleteOne({ _id: id }, (err, result) => {
+            if (!result) {
+              res.status(200).send({ message: err.message, success: false });
+            } else {
+              res.status(200).send({
+                message: "Data deleted Successfully",
+                success: true,
+                data: result,
+              });
+            }
           });
+        } else {
+          res.status(200).send({ message: err.message, success: false });
         }
       });
     }
@@ -122,7 +119,6 @@ router.get("/", (req, res) => {
               message: "Data get Successfully",
               success: true,
               data: result,
-              token: tokengenerate({ user: req.user }),
             });
           }
         }
@@ -136,7 +132,6 @@ router.get("/", (req, res) => {
             message: "Data get Successfully",
             success: true,
             data: result,
-            token: tokengenerate({ user: req.user }),
           });
         }
       });
